refactor: correção da estrutura do projeto
