+++ src/main/java/gestaofuncionarios/presenter/GestaoFuncionariosPresenter.java
@@ -6,8 +6,6 @@ import javax.swing.JInternalFrame;
 import javax.swing.JOptionPane;
 
 import gestaofuncionarios.dados.dao.FuncionarioDAO;
-import gestaofuncionarios.presenter.funcionariopresenter.FuncionarioPresenter;
-import gestaofuncionarios.presenter.funcionariopresenter.VerBonusPresenter;
 import gestaofuncionarios.view.GestaoFuncionariosView;
 
 /**
@@ -33,15 +31,6 @@ public class GestaoFuncionariosPresenter {
         this.view.getBuscarFuncionariosMenu().addActionListener((ActionEvent e) -> {
             try {
                 BuscarFuncionarioPresenter presenter = new BuscarFuncionarioPresenter(dao);
-                showPanel(presenter.getView(), false, false);
-            } catch (Exception ex) {
-                JOptionPane.showMessageDialog(view, ex.getMessage());
-            }
-        });
-
-        this.view.getShowHistoricoBonusMenu().addActionListener((ActionEvent e) -> {
-            try {
-                VerBonusPresenter presenter = new VerBonusPresenter();
                 showPanel(presenter.getView(), false, false);
             } catch (Exception ex) {
                 JOptionPane.showMessageDialog(view, ex.getMessage());
